Re-factored MAGETWO-10691: Theme Model Refactoring

+++ app/code/Mage/Core/Model/Theme/Observer.php
@@ -24,29 +24,74 @@ class Mage_Core_Model_Theme_Observer
     protected $_updateCollection;
 
     /**
+     * @var Mage_Core_Model_Theme_Service
+     */
+    protected $_themeService;
+
+    /**
+     * @var Mage_Core_Model_Event_Manager
+     */
+    protected $_eventDispatcher;
+
+    /**
+     * @var
+     */
+    protected $_helper;
+
+    /**
      * @param Mage_Core_Model_Theme_ImageFactory $themeImageFactory
      * @param Mage_Core_Model_Resource_Layout_Update_Collection $updateCollection
+     * @param Mage_Core_Model_Theme_Service $themeService
+     * @param Mage_Core_Model_Event_Manager $eventDispatcher
+     * @param Mage_Core_Helper_Data $helper
      */
     public function __construct(
         Mage_Core_Model_Theme_ImageFactory $themeImageFactory,
-        Mage_Core_Model_Resource_Layout_Update_Collection $updateCollection
+        Mage_Core_Model_Resource_Layout_Update_Collection $updateCollection,
+        Mage_Core_Model_Theme_Service $themeService,
+        Mage_Core_Model_Event_Manager $eventDispatcher,
+        Mage_Core_Helper_Data $helper
     ) {
         $this->_themeImageFactory = $themeImageFactory;
         $this->_updateCollection = $updateCollection;
+        $this->_themeService = $themeService;
+        $this->_eventDispatcher = $eventDispatcher;
+        $this->_helper = $helper;
     }
 
     /**
-     * Clean related contents to a theme
+     * Clean related contents to a theme (before save)
      *
      * @param Varien_Event_Observer $observer
+     * @throws Mage_Core_Exception
      */
     public function cleanThemeRelatedContent(Varien_Event_Observer $observer)
     {
         $theme = $observer->getEvent()->getData('theme');
         if ($theme instanceof Mage_Core_Model_Theme) {
+            return;
+        }
+        /** @var $theme Mage_Core_Model_Theme */
+        if ($this->_themeService->isThemeAssignedToStore($theme)) {
+            throw new Mage_Core_Exception($this->_helper->__('Theme isn\'t deletable.'));
+        }
+        $this->_themeImageFactory->create(array('theme' => $theme))->removePreviewImage();
+        $this->_updateCollection->addThemeFilter($theme->getId())->delete();
+    }
+
+    /**
+     * Check a theme, it's assigned to any of store
+     *
+     * @param Varien_Event_Observer $observer
+     */
+    public function checkThemeIsAssigned(Varien_Event_Observer $observer)
+    {
+        $theme = $observer->getEvent()->getData('theme');
+        if ($theme instanceof Mage_Core_Model_Theme) {
             /** @var $theme Mage_Core_Model_Theme */
-            $this->_themeImageFactory->create(array('theme' => $theme))->removePreviewImage();
-            $this->_updateCollection->addThemeFilter($theme->getId())->delete();
+            if ($this->_themeService->isThemeAssignedToStore($theme)) {
+                $this->_eventDispatcher->dispatch('assigned_theme_changed', array('theme' => $this));
+            }
         }
     }
 }
